refactor(meta): extract barrier scheduler from barrier manager
