+++ src/meta/src/rpc/service/stream_service.rs
@@ -22,7 +22,7 @@ use risingwave_pb::meta::stream_manager_service_server::StreamManagerService;
 use risingwave_pb::meta::*;
 use tonic::{Request, Response, Status};
 
-use crate::barrier::BarrierManagerRef;
+use crate::barrier::BarrierScheduler;
 use crate::manager::{FragmentManagerRef, MetaSrvEnv};
 use crate::storage::MetaStore;
 
@@ -34,7 +34,7 @@ where
     S: MetaStore,
 {
     env: MetaSrvEnv<S>,
-    barrier_manager: BarrierManagerRef<S>,
+    barrier_scheduler: BarrierScheduler<S>,
     fragment_manager: FragmentManagerRef<S>,
 }
 
@@ -44,12 +44,12 @@ where
 {
     pub fn new(
         env: MetaSrvEnv<S>,
-        barrier_manager: BarrierManagerRef<S>,
+        barrier_scheduler: BarrierScheduler<S>,
         fragment_manager: FragmentManagerRef<S>,
     ) -> Self {
         StreamServiceImpl {
             env,
-            barrier_manager,
+            barrier_scheduler,
             fragment_manager,
         }
     }
@@ -65,7 +65,7 @@ where
         self.env.idle_manager().record_activity();
         let _req = request.into_inner();
 
-        let snapshot = self.barrier_manager.flush().await?;
+        let snapshot = self.barrier_scheduler.flush().await?;
         Ok(Response::new(FlushResponse {
             status: None,
             snapshot: Some(snapshot),
