+++ src/meta/src/stream/scale.rs
@@ -684,7 +684,7 @@ where
                 .await;
         }));
 
-        self.barrier_manager
+        self.barrier_scheduler
             .run_multiple_commands(vec![
                 Command::Plain(Some(Mutation::Pause(PauseMutation {}))),
                 Command::RescheduleFragment(reschedule_fragment),
